requirements.txt
pandas==2.1.4
psycopg2==2.9.9
pyyaml==6.0.1
jupyter==1.0.0
ipykernel==6.29.0
schedule>=1.2.0
redis>=5.0.0
